requirements.txt
spacy>=3.8.0,<3.9.0
spacy-transformers
python-docx>=1.1.0
PyPDF2>=3.0.0
pillow